bug-fix: time range start time in unix can be negative if older than 1970

+++ src/api/dsl/Dsl.h
@@ -134,7 +134,7 @@ static LikeExpression like(const T& expr,
 template <typename T>
 static RoundTimeExpression round(const T& expr,
                                  size_t unit,
-                                 size_t startTime) {
+                                 int64_t startTime) {
   return RoundTimeExpression(std::shared_ptr<Expression>(new T(expr)),
                              unit,
                              startTime);
